feat: change folders structure
