+++ ImageSearcher/ImageSearcher/App/AppDelegate.swift
