--- ImageSearcher/ImageSearcher/MainViewController.swift
@@ -1,19 +1,0 @@
-//
-//  ViewController.swift
-//  ImageSearcher
-//
-//  Created by Kyzu on 28.11.22.
-//
-
-import UIKit
-
-class MainViewController: UIViewController {
-
-    override func viewDidLoad() {
-        super.viewDidLoad()
-    
-    }
-
-
-}
-
