--- ImageSearcher/ImageSearcher/SceneDelegate.swift
@@ -1,22 +1,0 @@
-//
-//  SceneDelegate.swift
-//  ImageSearcher
-//
-//  Created by Kyzu on 28.11.22.
-//
-
-import UIKit
-
-class SceneDelegate: UIResponder, UIWindowSceneDelegate {
-    
-    var window: UIWindow?
-    
-    func scene(_ scene: UIScene, willConnectTo session: UISceneSession, options connectionOptions: UIScene.ConnectionOptions) {
-        guard let windowScene = (scene as? UIWindowScene) else { return }
-        window = UIWindow(windowScene: windowScene)
-        window?.rootViewController = MainViewController()
-        window?.makeKeyAndVisible()
-    }
-    
-}
-
