+++ ImageSearcher/ImageSearcher/Views/BigPhotoViewController.swift
@@ -1,0 +1,8 @@
+//
+//  BigPhotoViewController.swift
+//  ImageSearcher
+//
+//  Created by Kyzu on 28.11.22.
+//
+
+import Foundation
