+++ ImageSearcher/ImageSearcher/Views/MainViewController.swift
@@ -1,0 +1,26 @@
+//
+//  ViewController.swift
+//  ImageSearcher
+//
+//  Created by Kyzu on 28.11.22.
+//
+
+import UIKit
+
+class MainViewController: UIViewController {
+
+    var viewModel = MainViewModel()
+    
+    override func viewDidLoad() {
+        super.viewDidLoad()
+        bindViewModel()
+    
+    }
+    
+    func bindViewModel() {
+        viewModel.statusText.bind(<#T##listener: Dynamic<String>.Listener?##Dynamic<String>.Listener?##(String) -> Void#>)
+    }
+
+
+}
+
